Add Methods on DecisionEngine and Create Metrics Module
I extracted the "observation type" type definitions into a module
called "metrics" to more closely reflect that the observations
could be other types of metrics, like CPU and RAM usage, in addition
to response codes.

I also added a little more detail to the trait definition of the
decision engine.